+++ packages/popmotion/src/action/_tests/vector.test.ts
@@ -1,5 +1,6 @@
 import { number } from 'style-value-types';
 import action from '../';
+import tween from '../../animations/tween';
 import vector from '../vector';
 
 describe('vector', () => {
@@ -31,5 +32,56 @@ describe('vector', () => {
             : reject('parallel output incorrect');
         });
     });
+  });
+
+  it('should return a composite action if provided a color', () => {
+    return new Promise((resolve, reject) => {
+      let col = '';
+      tween({
+        from: '#ff0',
+        to: 'rgba(0, 0, 0, 1)'
+      }).start({
+        update: (v) => col = v,
+        complete: () => (col === 'rgba(0, 0, 0, 1)') ? resolve() : reject()
+      });
+    });
+  });
+
+  it('should return a composite action if provided a color', () => {
+    return new Promise((resolve, reject) => {
+      let output = {};
+      tween({
+        from: {
+          x: 0,
+          background: '#ff0'
+        },
+        to: {
+          x: 100,
+          background: 'rgba(0, 0, 0, 1)'
+        }
+      }).start({
+        update: (v) => output = v,
+        complete: () => {
+          if (output.background === 'rgba(0, 0, 0, 1)' && output.x === 100) {
+            resolve();
+          } else {
+            reject(`${output.background}, ${output.x}`);
+          }
+        }
+      });
+    });
+
+    it('should support hsla', () => {
+      return new Promise((resolve, reject) => {
+        let col = '';
+        tween({
+          from: 'hsla(0, 90, 10, 1)',
+          to: 'hsla(100, 90, 10, 0.5)'
+        }).start({
+          update: (v) => col = v,
+          complete: () => (col === 'hsla(100, 90, 10, 0.5)') ? resolve() : reject()
+        });
+      })
+    })
   });
 });